requirements.txt
###### Requirements without Version Specifiers ######


###### Requirements with Version Specifiers ######
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
requests==2.32.3
requests-oauthlib==2.0.0
yt-dlp==2024.11.18
